[ADD] working 3d

+++ include/display/window.h
@@ -17,4 +17,4 @@ void window_destroy(window_t *window);
 void window_clear(const window_t *window, const color_t *color);
 void window_update(const window_t *window);
 
-void window_draw_line(const window_t *window, int x1, int y1, int x2, int y2, const color_t *color);
+void window_draw_line(const window_t *window, float x1, float y1, float x2, float y2, const color_t *color);
